I... fixed a bug?
I haven't worked on this repository for a while, but there are quite a few changes here. Besides the obvious fancy renders, it looks like I may have fixed some kind of bug. I guess I also changed up the pole, probably to make That Continent Below the One With Militaryland on It look nicer.